pyproject.toml
[tool.poetry]
name = "pumpreels-bot"
version = "0.1.0"
description = ""
authors = ["Grossyb <grossyb@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.12"
fastapi = "^0.115.12"
requests = "^2.32.3"
uvicorn = "^0.34.0"
aiofiles = "^24.1.0"
python-telegram-bot = "^22.0"
google-cloud-storage = "^3.1.0"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
